+++ internal/prediction/holtwinters/holtwinters_test.go
@@ -440,7 +440,7 @@ func TestPredict_GetPrediction(t *testing.T) {
 			errors.New("holt winters algorithm error"),
 			&holtwinters.Predict{
 				Runner: &fake.Run{
-					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int64) (string, error) {
+					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int) (string, error) {
 						return "", errors.New("holt winters algorithm error")
 					},
 				},
@@ -505,7 +505,7 @@ func TestPredict_GetPrediction(t *testing.T) {
 			errors.New(`strconv.Atoi: parsing "invalid": invalid syntax`),
 			&holtwinters.Predict{
 				Runner: &fake.Run{
-					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int64) (string, error) {
+					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int) (string, error) {
 						return "invalid", nil
 					},
 				},
@@ -570,7 +570,7 @@ func TestPredict_GetPrediction(t *testing.T) {
 			nil,
 			&holtwinters.Predict{
 				Runner: &fake.Run{
-					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int64) (string, error) {
+					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int) (string, error) {
 						return "0", nil
 					},
 				},
@@ -647,7 +647,7 @@ func TestPredict_GetPrediction(t *testing.T) {
 			nil,
 			&holtwinters.Predict{
 				Runner: &fake.Run{
-					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int64) (string, error) {
+					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int) (string, error) {
 						return "2", nil
 					},
 				},
@@ -721,7 +721,7 @@ func TestPredict_GetPrediction(t *testing.T) {
 			nil,
 			&holtwinters.Predict{
 				Runner: &fake.Run{
-					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int64) (string, error) {
+					RunAlgorithmWithValueReactor: func(algorithmPath, value string, timeout int) (string, error) {
 						return "3", nil
 					},
 				},
